Добавлен модуль обработчиков окна полноразмерного фото
