+++ js/full-photo.js
@@ -1,4 +1,6 @@
-const fullPhoto = (photos) => {
+import { modalOpen } from './photo-modal.js';
+
+const openFullPhoto = (photos) => {
   const fullPhotoModal = document.querySelector('.big-picture');
   const thumbnails = document.querySelectorAll('.picture');
   const modalPicture = fullPhotoModal.querySelector('.big-picture__img').querySelector('img');
@@ -7,13 +9,10 @@ const fullPhoto = (photos) => {
   const commentsList = fullPhotoModal.querySelector('.social__comments');
   const commentTemplate = commentsList.querySelector('.social__comment');
   const photoDescription = document.querySelector('.social__caption');
-  const commentsCountModal = document.querySelector('.social__comment-count');
-  const commentsLoaderButton = document.querySelector('.comments-loader');
-  const modalCloseButton = document.querySelector('#picture-cancel');
 
   const addThumbnailHandler = (thumbnail, photo) => {
     thumbnail.addEventListener ('click', () => {
-      fullPhotoModal.classList.remove('hidden');
+      modalOpen();
       modalPicture.src = photo.url;
       likes.textContent = photo.likes;
       commentsCount.textContent = photo.comments.length;
@@ -34,32 +33,12 @@ const fullPhoto = (photos) => {
       });
 
       commentsList.append(commentsListFragment);
-
-      commentsCountModal.classList.add('hidden');
-      commentsLoaderButton.classList.add('hidden');
-      document.body.classList.add('modal-open');
     });
   };
 
   for (let i = 0; i < thumbnails.length; i++) {
     addThumbnailHandler(thumbnails[i], photos[i]);
   }
-
-  modalCloseButton.addEventListener('click', () => {
-    fullPhotoModal.classList.add('hidden');
-    commentsCountModal.classList.remove('hidden');
-    commentsLoaderButton.classList.remove('hidden');
-    document.body.classList.remove('modal-open');
-  });
-
-  document.addEventListener('keydown', (evt) => {
-    if (evt.key === 'Escape') {
-      fullPhotoModal.classList.add('hidden');
-      commentsCountModal.classList.remove('hidden');
-      commentsLoaderButton.classList.remove('hidden');
-      document.body.classList.remove('modal-open');
-    }
-  });
 };
 
-export {fullPhoto};
+export {openFullPhoto};
